pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Alex Green <alexgreendev@gmail.com>"]
readme = "README.md"
repository = "https://github.com/alexgreendev/python-project-52.git"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
packages = [
    {include = "task_manager"}
]


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^4.2.8"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
gunicorn = "^21.2.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
django-bootstrap5 = "^23.3"
coverage = "^7.3.2"
django-extensions = "^3.2.3"
django-filter = "^23.4"
rollbar = ">=0.16,<1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
